Break out client-navigation and rendering test from basic test
