+++ test/integration/client-navigation/components/hello1.js
@@ -1,0 +1,3 @@
+export default () => (
+  <p>Hello World 1</p>
+)
