+++ test/integration/client-navigation/components/world.jsx
