+++ test/integration/client-navigation/lib/async-function.js
