+++ test/integration/client-navigation/pages/async-props.js
