+++ test/integration/client-navigation/pages/circular-json-error.js
