+++ test/integration/client-navigation/pages/custom-encoding.js
