+++ test/integration/client-navigation/pages/custom-extension.jsx
