+++ test/integration/client-navigation/pages/default-head.js
