+++ test/integration/client-navigation/pages/dynamic/ssr.js
@@ -1,0 +1,5 @@
+import dynamic from 'next/dynamic'
+
+const Hello = dynamic(import('../../components/hello1'))
+
+export default Hello
