+++ test/integration/client-navigation/pages/error-in-the-browser-global-scope.js
