+++ test/integration/client-navigation/pages/error-in-the-global-scope.js
