+++ test/integration/client-navigation/pages/error-inside-browser-page.js
