+++ test/integration/client-navigation/pages/error-inside-page.js
