+++ test/integration/client-navigation/pages/exports.js
