+++ test/integration/client-navigation/pages/forwardRef-component.js
