+++ test/integration/client-navigation/pages/fragment-syntax.js
