+++ test/integration/client-navigation/pages/head.js
