+++ test/integration/client-navigation/pages/index.js
