+++ test/integration/client-navigation/pages/instance-get-initial-props.js
