+++ test/integration/client-navigation/pages/json.js
