+++ test/integration/client-navigation/pages/link.js
