+++ test/integration/client-navigation/pages/memo-component.js
