+++ test/integration/client-navigation/pages/nav/as-path-pushstate.js
