+++ test/integration/client-navigation/pages/nav/as-path-query.js
