+++ test/integration/client-navigation/pages/nav/as-path-using-router.js
