+++ test/integration/client-navigation/pages/nav/as-path.js
