+++ test/integration/client-navigation/pages/nav/hash-changes.js
