+++ test/integration/client-navigation/pages/nav/index.js
