+++ test/integration/client-navigation/pages/nav/on-click.js
