+++ test/integration/client-navigation/pages/nav/querystring.js
