+++ test/integration/client-navigation/pages/nav/redirect.js
