+++ test/integration/client-navigation/pages/nav/self-reload.js
