+++ test/integration/client-navigation/pages/nav/shallow-routing.js
