+++ test/integration/client-navigation/pages/nav/with-hoc.js
