+++ test/integration/client-navigation/pages/nested-cdm/index.js
