+++ test/integration/client-navigation/pages/nested-index/index/index.js
