+++ test/integration/client-navigation/pages/no-default-export.js
