+++ test/integration/client-navigation/pages/read-only-object-error.js
