+++ test/integration/client-navigation/pages/stateless.js
