+++ test/integration/client-navigation/pages/styled-jsx.js
