+++ test/integration/client-navigation/pages/throw-undefined.js
