+++ test/integration/client-navigation/pages/with-cdm.js
